Prevent retry script from failing when cleanup fails (#88)
*Issue #, if available:*
This task was started because it was believed that the cleanup command
had a bug that caused the retry scripts to exit prematurely when failed.
However, testing shows that the retry scripts works fine.

- Test
[run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9473507110/job/26101288040)
with failure: Left out the change and added "exit 1" to the actual and
cleanup command on the Download Enablement Script Step to verify that
the script exits prematurely. It did not.

- Test
[run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9473637165)
with failure: Same as above but with bug fixed. Verify that the retry
worked 3 times

An example
[run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/8935086952/job/24543026873)
where the cleanup command failed shows that it was during the Terraform
Deployment step, which uses the retry script and not the action. The
retry script doesn't seem to have the cleanup fail prevention
implemented, hence why this run exited prematurely.

Retry Action Script has a small bug that makes it such that failure in
cleanup command will make retry exit prematurely.

*Description of changes:*
Add the cleanup prevention step to the retry script

Test
[run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9489969220)
with success

Test
[run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9490098736)
where cleanup didn't exit prematurely for action

Test
[run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9490300250)
where cleanup didn't exit prematurely for script



By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.