Increase Retry Wait Times and Add Additional Gradlew Step (#90)
*Issue #, if available:*
Increasing the wait times for `Terraform Init`, and `Gradlew` Step to
mitigate the occurrence of these issues failing, especially `Terraform
Init`.

For Gradlew, wait time is 60 and retry is 3. If it fails initially, will
try again with the same parameters. If retry triggers, it will add a max
of 6 minutes.

For Terraform Init, wait time is 60 and retry is 6. If retry triggers,
it will add a max of 6 minutes

Most runs are 8-10 minutes without the retries, so these additional
retries could cause the runs to last more than 15 minutes. Since it's
rare for these errors to occur consecutively, should not be an issue if
it exceeds 15 minutes by a minute or two.

Additionally, adding `Initiate Gradlew Step` to the beginning of the run
so that if it fails, we can try it again before the validator step.
Cannot do the same for `Terraform Init` because this steps already
occurs very early on during the run.

The task also involved increasing wait time for the enable-app-signal.sh
script, but since the cleanup command wasn't working and is fixed on
this
[PR](https://github.com/aws-observability/aws-application-signals-test-framework/pull/88),
will leave it as it is for now.


Test run:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9489300435
Test run after adding continue-on error:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/9508162874


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.