Propagate RDS database endpoint and credentials to the SampleApp and expose a new API to query the database (#96)
*Issue description:*
in Application Signals E2E tests, we are not testing database related
span today.

*Description of changes:*
This change will introduce two new Github workflow steps for Java E2E
canary to call RDS to get the endpoint and cluster identifier via CLI,
store them as Github environment variables, and then use the cluster
identifier to call SecretsManager via CLI to get the secrets name. In
the second workflow step, it uses secrets manager action to retrieve
credentials based on the secrets name for RDS cluster obtained from the
previous step.

Java Sample APP is also updated to expose a new API: `mysql` which makes
a simple query: `select * from tables limit 1;` from the
`information_schema` database which is a informational built-in database
for all MySQL databases;

Invoking `/mysql` API and asserting desired span information will be
implemented in a separate PR. Same for Python.

This has been tested here:
https://github.com/jerry-shao/aws-application-signals-test-framework/actions/runs/9700780569/job/26772985493

We can safely merge this PR even the RDS database is not available as we
are not using it yet. See E2E test run when I use `mydatabase` as the
database name. The workflow continues on these steps when we get error
from RDS saying the cluster was not found.

https://github.com/jerry-shao/aws-application-signals-test-framework/actions/runs/9764329486/job/26952410687

*Ensure you've run the following tests on your changes and include the
link below:*
To do so, create a `test.yml` file with `name: Test` and workflow
description to test your changes, then remove the file for your PR. Link
your test run in your PR description. This process is a short term
solution while we work on creating a staging environment for testing.

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.