[Python] Add ec2-requirements.txt for otel version upgrade (#108)
*Issue description:*
ADOT Python version is doing an OTEL version upgrade to 1.25 that
conflicts with this E2E test ONLY for EC2. OTEL version 1.22 is
currently being used by the sample app to get the trace ID of the span
sent by the instrumentation for the validator to validate the generated
trace.

*Description of changes:*
- Create a new ec2-requirements.txt file only for the EC2 use case

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.