[All] Refresh credentials before emitting metrics and cleanup (#138)
*Issue description:*
The total length of all of our retries exceeds 1 hour in the worst case,
especially if we face a transient set up issue in terraform. This change
will ensure that credentials are still available before the metric
publication and clean up is done so that the next test run can function
appropriately.