Metric emission v2 (#151)
*Issue description:*
This PR aims to resolve two primary issues during e2e testing:
1. E2E tests lasting more than 1 hour results in AWS credentials
expiring and result not being published
2. E2E test failing to start due to issues such as ECR throttling while
trying to start container

*Description of changes:*
We will have a new job that is responsible for publishing the success
metric after the e2e job is finished. Each e2e job run will output a
variable with either `success` or`failure` result. The publish metric
job will read those output and publish the appropriate result to
cloudwatch metric

*Rollback procedure:*
Check if the metrics are appearing in at least 2 regions and if not,
revert this PR

*Test run:*

https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10220378559/job/28281118016

*Failed test run*

https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10219823497


First run to check if success metrics arrived to cloudwatch metric
Second run to check if fail metric arrive when failed.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.