Update env variable in publish-result workflow (#156)
*Issue description:*
Using the wrong secret in e2e-test-publish-result.yml. Although this is
working in the framework repository, it is causing failure in release
testing repos because the secret is not set there.

Example failed run:
https://github.com/aws/amazon-cloudwatch-agent-operator/actions/runs/10222921603/job/28288546373

*Description of changes:*
Update the secret to an env variable instead

*Rollback procedure:*
Revert the PR

Test run:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10223050553

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.