Check if sample app is up (#162)
*Issue description:*
After removing the public endpoints, we removed the logic for checking
if the sample application is up since we are unable to ping the sample
app through the workflow. However, recently there have been increasing
failures in EC2s due to the sample application not generating telemetry.
This PR aims to resolve this issue by adding the logic back

*Description of changes:*
Check if the sample application is up after building it in Terraform. If
the sample application is unavailable, then throw exit 1 to indicate
that the Terraform Deployment was a failure. This will trigger the e2e
workflow to re-instantiate another EC2 instance and deploy the sample
application again.

For ASGs, since the ASG creates the EC2 instance we need a different way
to check if the sample application is available or not since throwing
exit 1 will not cause a failure. Will push this PR first then apply
implementation for ASG if we still noticeable improvement

*Rollback procedure:*
Revert PR

Test run:
-
[Run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10291998960)
to check if EC2s will pass after new implementation:
-
[Run](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10291376482/job/28483485792)to
check if Terraform will fail and retry if sample application is not
available (Did this by changing the port to an invalid one)

Test run after updating:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10292445294/job/28486870614

Final Test run:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10309116932/job/28538036319


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.