[EKS] Remove `/mysql` temporarily
Checking the following test run, found that the only telemetry generated was for `/mysql`. This means that this test case may be impacting the ability of the sample app to move on to other endpoints. Since we don't test this case yet, removing it to see if there's an improvement to be seen

https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10333263032/job/28605408100