Remove sample app namespace in the beginning (#182)
*Issue description:*
We had an issue where previous sample applications did not get deleted
because a runner abruptly disconnected in the middle of the run.
([Link](https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10437137885/job/28902975541))
This caused subsequent EKS runs to fail due to the port being occupied
and had to be manually cleaned

*Description of changes:*
Clean up the EKS cluster namespace before deploying a new sample
application

Test run:

https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10532509683/job/29186804244
Also manually tried running the command

*Rollback procedure:*
Revert the PR

<Can we safely revert this commit if needed? If not, detail what must be
done to safely revert and why it is needed.>

*Ensure you've run the following tests on your changes and include the
link below:*

To do so, create a `test.yml` file with `name: Test` and workflow
description to test your changes, then remove the file for your PR. Link
your test run in your PR description. This process is a short term
solution while we work on creating a staging environment for testing.

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.