Patch K8s on EC2 OS Part 1 (#191)
*Issue description:*
The EC2 AMI has an expiry date. Since our K8s on EC2 are designed to
last indefinitely, eventually we will have the update the AMI. This PR
aims to automate updating the AMIs without any manual interference. This
will involve creating a new EC2 instance and taking down the old one

[Part
1](https://github.com/aws-observability/aws-application-signals-test-framework/pull/191):
Create a new K8s on EC2 and test it
[Part
2](https://github.com/aws-observability/aws-application-signals-test-framework/compare/k8s-automation-part-2?expand=1):
Verify that the instances were created properly and no impact on
existing K8s, then replace the K8s

*Description of changes:*
**k8s-patch-os-matrix.yml**
Define all K8s on EC2s instances including the ones for canary as well
as release testing. Any new K8s on EC2s will be defined here. It will
update the AMI every month

**k8s-patch-os-jobs.yml**
1. Run a script that was copied and modified from
[here](https://github.com/aws-observability/application-signals-demo/blob/main/scripts/k8s/appsignals/setup-k8s-demo.sh).
Creates new key pair and two ec2 instances, which will have kubernetes
installed
2. Store the newly created public ip and private ssh key in a secret key
called `-pending-value`. Do not store it in the main secret yet because
we want to make sure that our K8s test passes before updating the main
secret key.
3. Run K8s E2E test on newly created instance and verify it succeeds.
All tests have been figured to use the release artifacts, including the
ones for release testing since we don't want the test to fail due to
invalid staging artifacts
4. Save main secret values onto a key called `-temporary-storage` then
replace the main secret values with the newly created ones. This is so
that in case of issues with the new k8s, we can find the original values
and revert it back

Testing Id for K8s was updated to give a random value because all the
K8s tests run under the same run id and job id is not retrievable in
Github Action. Needed a way to distinguish between different runs or the
validators will not be able to tell which telemetry to retrieve

Updated the ssm parameter key to use testing id instead of repository to
make it unique to each run

*Rollback procedure:*
Revert PR, newly created EC2 instances will not affect canary. Make sure
that the main secret keys were not altered.

*Test*
Test run:
https://github.com/harrryr/aws-application-signals-test-framework/actions/runs/10574639963
Test run on canary k8s:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/10593567350

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.