Patch K8s on EC2 OS Part 2 (#192)
*Issue description:*
The EC2 AMI has an expiry date. Since our K8s on EC2 are designed to
last indefinitely, eventually we will have the update the AMI. This PR
aims to automate updating the AMIs without any manual interference. This
will involve creating a new EC2 instance and taking down the old one

[Part
1](https://github.com/aws-observability/aws-application-signals-test-framework/pull/191):
Create a new K8s on EC2 and test it
[Part
2](https://github.com/aws-observability/aws-application-signals-test-framework/compare/k8s-automation-part-2?expand=1):
Verify that the instances were created properly and no impact on
existing K8s, then replace the K8s

Things to check:
Verify that the canary K8s are stable
Verify that new instance was created and the secret in `-pending-value`
are correct

*Description of changes:*
Uncomment section for replacing main secret keys with the newly created
ones.

*Rollback procedure:*
Revert, move the value in `-temporary-storage` back to the the main
values

To do so, create a `test.yml` file with `name: Test` and workflow
description to test your changes, then remove the file for your PR. Link
your test run in your PR description. This process is a short term
solution while we work on creating a staging environment for testing.

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.