Added Dotnet Lambda E2E Tests (#332)
*Description of changes:*
This change adds workflow files and terraform scripts that builds the
infrastructure (lambda function and lambda layer) for DotNet lambda
layer. For now, the cron schedule in
[dotnet-lambda-canary.yml](https://github.com/aws-observability/aws-application-signals-test-framework/compare/dotnetLambdaE2E?expand=1#diff-7f7da6fc2394c58665b91c661888dfc9afec0edf64c1a53701e56a7fcdf943bd)
is disabled which means that tests won't run yet. We will need to wait
until we officially release the lambda layer. However, I updated the
aws-otel-dotnet-instrumentation build workflow to run the lambda tests
on main build/merge.

Testing:
Check this run for details, especially dotnet-lambda-default step:
https://github.com/aws-observability/aws-otel-dotnet-instrumentation/actions/runs/12283246401/job/34276401374.
This verifies that E2E tests work and that it runs as part of the main
build for the `aws-otel-dotnet-instrumentation`. (I disabled other E2E
tests for the run above just to save time.)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.