update remote deployment names  (#370)
*Issue description:*

*Description of changes:*

Shorten the deployment names for k8s and add a Kubernete service for
each remote app deployment. This is related to some behavior changes for
cloudwatch agent:

*
https://github.com/aws/amazon-cloudwatch-agent/commit/fc535d172b0874ad9b05361d570adc72af170eff
*
https://github.com/aws/amazon-cloudwatch-agent/commit/c53bdcbd0ba183317df14b7c161058cbeb45c5bb

**New behavior:** 
* we will only do ip to workload mapping if a workload is associated
with a kubernetes service
* if podname doesn't follow regular regex pattern, we will use service
name (rather than workload name) as RemoteService

In addition, modify the deployment name in dotnet eks windows test case
that I forgot in last PR.

*Rollback procedure:*

<Can we safely revert this commit if needed? If not, detail what must be
done to safely revert and why it is needed.>

*Ensure you've run the following tests on your changes and include the
link below:*

To do so, create a `test.yml` file with `name: Test` and workflow
description to test your changes, then remove the file for your PR. Link
your test run in your PR description. This process is a short term
solution while we work on creating a staging environment for testing.

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.