Update yq command to work on new yaml (#375)
*Issue description:*
Due to recent
[change](https://github.com/aws-observability/aws-application-signals-test-framework/commit/5392eb23c0d77a2bfbaf82561f34c9aec722dd1f#diff-1d6967a449c1950768de88fc7005718c165b077f045bdbd7d30348f3dcb66e4c)
with the yaml file for k8s, the release testing workflow is failing.

*Description of changes:*
Update the yq command to work with the updated yaml file

*Testing*:
Ran command locally, verified that it updated appropriately

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.