[Java] Remove EKS/K8s/ECS/EC2 canaries (#382)
## Background/Changes

The EKS/K8s/ECS/EC2 canaries are being deprecated in favour of
internally run tests. The test workflows will be maintained to be used
for release testing, but the retry/canary workflows will be removed now
as their alarms have been removed.

Java EKS OTLP OCB canary is being kept alongside the Lambda canary for
which we have not yet performed a migration.

## Rollback procedure

Revert

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.