[.NET] Remove EKS/K8s/EC2 canaries (#385)
## Background/Changes

The EKS/K8s/EC2 canaries are being deprecated in favour of internally
run tests. The test workflows will be maintained to be used for release
testing, but the retry/canary workflows will be removed now as their
alarms have been removed.

The lambda canary will be kept as we have not yet created replacement
tests for those.

## Rollback procedure

Revert

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.