Fix a typo that stop CWAgent release test from working (#405)
*Issue description:*
Fix a typo that stop CWAgent release test from working

Noted that in current config,
https://github.com/aws-observability/aws-application-signals-test-framework/blob/759916951597246bb20a626693fb506ace69344c/sample-apps/dotnet/dotnet-ec2-win-main-setup.ps1#L25

Test script require that the CWA to be named to
amazon-cloudwatcih-agent.msi for the script to run, the staging image
grab by s3 from test repo is NOT match this naming method, thus CWA is
not running in the first place, result in the test not working

*Test*

https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/15052116489/job/42309601468

*Description of changes:*
rename S3 download target name for CWAgent release test to match normal
cases
*Rollback procedure:*
revert
<Can we safely revert this commit if needed? If not, detail what must be
done to safely revert and why it is needed.>

*Ensure you've run the following tests on your changes and include the
link below:*

To do so, create a `test.yml` file with `name: Test` and workflow
description to test your changes, then remove the file for your PR. Link
your test run in your PR description. This process is a short term
solution while we work on creating a staging environment for testing.

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.