[Python] Renaming files and folders associated with Python SigV4 logs Release Testing (#414)
*Description of changes:*
The changes are made to be consistent with the Java SigV4 Logs Release
Testing-
[https://github.com/aws-observability/aws-application-signals-test-framework/pull/412](url)

Test workflow link
-https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/15571938763/job/43849467481

1. Simplified file paths by:

- Removing redundant 'traces' directory
- Standardizing 'adot-aws-otlp' paths
- Renaming validation files for clarity

2. Updated workflow configurations and job identifiers to match new
structure

3. Refactored template file paths in PredefinedExpectedTemplate.java
for:

- AWS SDK calls
- Remote service calls
- Client calls
- Updated language-specific documentation comments

4. Renamed multiple .mustache files to align with new structure (content
unchanged)

*Rollback procedure:*

<Can we safely revert this commit if needed? If not, detail what must be
done to safely revert and why it is needed.>

*Ensure you've run the following tests on your changes and include the
link below:*

To do so, create a `test.yml` file with `name: Test` and workflow
description to test your changes, then remove the file for your PR. Link
your test run in your PR description. This process is a short term
solution while we work on creating a staging environment for testing.

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Jeel Mehta <jeelmm@amazon.com>